Group RDAP integration tests
